:truck: moved link_ingest to its own file
